STL modernizations / libc++ support for types (OSX)
